Move sha computation to a dependency specific utils package
