+++ smithy-cli/src/test/resources/software/amazon/smithy/cli/dependencies/sha-test.txt
