+++ smithy-utils/src/main/java/software/amazon/smithy/utils/IoUtils.java
@@ -31,9 +31,6 @@ import java.nio.file.Paths;
 import java.nio.file.SimpleFileVisitor;
 import java.nio.file.StandardCopyOption;
 import java.nio.file.attribute.BasicFileAttributes;
-import java.security.DigestInputStream;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
@@ -366,40 +363,6 @@ public final class IoUtils {
         }
 
         return true;
-    }
-
-    /**
-     * Computes the sha1 digest of a file.
-     *
-     * @param path Path to file to compute hash for.
-     * @return sha1 digest string.
-     * @throws UncheckedIOException if the specified file could not be read.
-     */
-    public static String computeSha1(Path path) {
-        try (InputStream in = Files.newInputStream(path)) {
-            MessageDigest md = MessageDigest.getInstance("SHA-1");
-            try (DigestInputStream din = new DigestInputStream(in, md)) {
-                byte[] buf = new byte[1024 * 32];
-                int n;
-                do {
-                    n = din.read(buf);
-                } while (n > 0);
-            }
-            StringBuilder sb = new StringBuilder();
-            for (byte b : md.digest()) {
-                int decimal = (int) b & 0xff;
-                String hex = Integer.toHexString(decimal);
-                if (hex.length() == 1) {
-                    sb.append('0');
-                }
-                sb.append(hex);
-            }
-            return sb.toString();
-        } catch (NoSuchAlgorithmException e) {
-            throw new RuntimeException(e);
-        } catch (IOException e) {
-                throw new UncheckedIOException(e);
-        }
     }
 
     public static void copyDir(Path src, Path dest) {
