+++ smithy-utils/src/test/java/software/amazon/smithy/utils/IoUtilsTest.java
@@ -30,7 +30,6 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Objects;
 import java.util.Random;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -148,14 +147,5 @@ public class IoUtilsTest {
         Assertions.assertThrows(IllegalArgumentException.class, () -> IoUtils.rmdir(path));
 
         Files.delete(path);
-    }
-
-    @Test
-    public void computesCorrectShaHash() throws URISyntaxException {
-        // Expected sha was computed using the `sha1sum` command line utility
-        String expectedSha = "543d822a441db77f32309650a1b3c1510c9be392";
-        String shaHash = IoUtils.computeSha1(Paths.get(Objects.requireNonNull(
-                getClass().getResource("sha-test.txt")).toURI()));
-        assertEquals(shaHash, expectedSha);
     }
 }
